Refactor extract git.CommitTemplate
The CoAuthor concept is specific to Git commit templates and
does make little sense for other kinds of templates.

I believe template.T could be used to template other files (Git configuration
files for example) and would better be kept generic.
